+++ pineappl_cli/src/import.rs
@@ -139,7 +139,7 @@ impl Subcommand for Opts {
             println!("can not check conversion for this type");
         } else {
             let mut pdf = helpers::create_pdf(&self.pdfset)?;
-            let results = helpers::convolute(&grid, &mut pdf, &[], &[], &[], 1, false);
+            let results = helpers::convolute(&grid, &mut pdf, &[], &[], &[], 1, false, false);
 
             let mut table = helpers::create_table();
             table.set_titles(row![c => "b", "PineAPPL", grid_type, "rel. diff"]);